change toolkit to accept custom client
