+++ packages/flex-plugins-api-client/src/clients/__tests__/client.test.ts
@@ -100,7 +100,7 @@ describe('PluginServiceHttp', () => {
 
     it('should create no query parameter if no pagination is provided', async () => {
       get.mockResolvedValue({ meta: {}, data: [] });
-      const result = await client.list('/the-url');
+      const result = await client.list('/the-url', 'data');
 
       expect(result).toEqual({ meta: {}, data: [] });
       expect(get).toHaveBeenCalledTimes(1);
@@ -109,7 +109,7 @@ describe('PluginServiceHttp', () => {
 
     it('should add one pagination parameter', async () => {
       get.mockResolvedValue({ meta: {}, data: [] });
-      const result = await client.list('/the-url', { pageSize: 5 });
+      const result = await client.list('/the-url', 'data', { pageSize: 5 });
 
       expect(result).toEqual({ meta: {}, data: [] });
       expect(get).toHaveBeenCalledTimes(1);
@@ -118,7 +118,7 @@ describe('PluginServiceHttp', () => {
 
     it('should add multiple pagination parameters', async () => {
       get.mockResolvedValue({ meta: {}, data: [] });
-      const result = await client.list('/the-url', { page: 1, pageSize: 5 });
+      const result = await client.list('/the-url', 'data', { page: 1, pageSize: 5 });
 
       expect(result).toEqual({ meta: {}, data: [] });
       expect(get).toHaveBeenCalledTimes(1);
@@ -132,7 +132,7 @@ describe('PluginServiceHttp', () => {
         },
         data: [],
       });
-      const result = await client.list('/the-url');
+      const result = await client.list('/the-url', 'data');
 
       expect(result.meta.next_token).toEqual('123');
       expect(get).toHaveBeenCalledTimes(1);
@@ -146,11 +146,27 @@ describe('PluginServiceHttp', () => {
         },
         data: [],
       });
-      const result = await client.list('/the-url');
+      const result = await client.list('/the-url', 'data');
 
       expect(result.meta.previous_token).toEqual('321');
       expect(get).toHaveBeenCalledTimes(1);
       expect(get).toHaveBeenCalledWith('/the-url?');
+    });
+
+    it('should change response key with provided key', async () => {
+      get.mockResolvedValue({
+        meta: {
+          previous_token: '321',
+        },
+        results: [],
+      });
+      const result = await client.list('/the-url', 'plugins');
+
+      expect(result.meta.previous_token).toEqual('321');
+      expect(get).toHaveBeenCalledTimes(1);
+      expect(get).toHaveBeenCalledWith('/the-url?');
+      // @ts-ignore
+      expect(result.plugins).toEqual([]);
     });
   });
 });
