+++ packages/flex-plugins-api-client/src/clients/__tests__/configuredPluginsClient.test.ts
@@ -4,6 +4,7 @@ import PluginServiceHttpClient from '../client';
 describe('ConfiguredPluginsClient', () => {
   const httpClient = new PluginServiceHttpClient('username', 'password');
   const get = jest.spyOn(httpClient, 'get');
+  const list = jest.spyOn(httpClient, 'list');
   const client = new ConfiguredPluginsClient(httpClient);
 
   beforeEach(() => {
@@ -11,13 +12,14 @@ describe('ConfiguredPluginsClient', () => {
   });
 
   it('should list configured plugins', async () => {
-    get.mockResolvedValue('list');
+    // @ts-ignore
+    list.mockResolvedValue({ meta: {}, data: {} });
 
     const result = await client.list('configId');
 
-    expect(result).toEqual('list');
-    expect(get).toHaveBeenCalledTimes(1);
-    expect(get).toHaveBeenCalledWith('Configurations/configId/Plugins', { cacheable: true });
+    expect(result).toEqual({ meta: {}, data: {} });
+    expect(list).toHaveBeenCalledTimes(1);
+    expect(list).toHaveBeenCalledWith('Configurations/configId/Plugins', 'plugins');
   });
 
   it('should get configured plugin', async () => {
